// evm/types/contracts/interfaces/IHashflowRouter.ts
export declare namespace IQuote {
// ...
  export type XChainRFQTQuoteStruct = {
    srcChainId: BigNumberish;
    dstChainId: BigNumberish;
    srcPool: AddressLike;
    dstPool: BytesLike;
    srcExternalAccount: AddressLike;
    dstExternalAccount: BytesLike;
    dstTrader: BytesLike;
    baseToken: AddressLike;
    quoteToken: BytesLike;
    effectiveBaseTokenAmount: BigNumberish;
    baseTokenAmount: BigNumberish;
    quoteTokenAmount: BigNumberish;
    quoteExpiry: BigNumberish;
    nonce: BigNumberish;
    txid: BytesLike;
    xChainMessenger: AddressLike;
    signature: BytesLike;
  };
// ...
}